requirements.txt
scikit-learn>=1.3.0
numpy>=1.24.0
supabase>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
requests>=2.31.0
slowapi>=0.1.9
joblib>=1.3.0
pytest>=8.0.0
httpx>=0.27.0
